bring back getGasPrice to controller/main.ts as it doesn't work otherwise
